refactor/EV-804/comentários apagados
