+++ lib/presenter/models/coins.dart
